made sending player data easier
Before, you had to manually specify which fields of the player you wanted to send for a given message, which is kind of annoying. Now, you can just call toJSON() on a player object to get a JSON-ified version of the player object that can be easily queried on the frontend easily.